requirement.txt
pillow
tensorflow
matplotlib
scikit-learn
cairosvg